Don't try running in Docker if there is no Docker

+++ Makefile
@@ -3,6 +3,7 @@ SHELL=bash
 
 .PHONY: *
 
+DOCKER_AVAILABLE=$((command -v docker >/dev/null 2>&1; echo $$?)
 CONTAINER_REGISTRY_REPO="ghcr.io/wyrihaximusnet/php"
 COMPOSER_SHOW_EXTENSION_LIST_PROD=$(shell (((command -v composer >/dev/null 2>&1) && composer show -t --no-plugins) || docker run --rm -v "`pwd`:`pwd`" -w `pwd` ${CONTAINER_REGISTRY_REPO}:8.4-nts-alpine-slim-dev composer show -t --no-plugins) | grep -o "\-\-\(ext-\).\+" | sort | uniq | cut -d- -f4- | tr -d '\n' | grep . | sed  '/^$$/d' | xargs | sed -e 's/ /, /g' | tr -cd '[:alnum:],' | sed 's/.$$//')
 COMPOSER_SHOW_EXTENSION_LIST_DEV=$(shell (((command -v composer >/dev/null 2>&1) && composer show -s --no-plugins) || docker run --rm -v "`pwd`:`pwd`" -w `pwd` ${CONTAINER_REGISTRY_REPO}:8.4-nts-alpine-slim-dev composer show -s --no-plugins) | grep -o "\(ext-\).\+" | sort | uniq | cut -d- -f2- | cut -d" " -f1 | xargs | sed -e 's/ /, /g' | tr -cd '[:alnum:],')
@@ -23,12 +24,16 @@ endif
 ifeq ("$(IN_DOCKER)","TRUE")
 	DOCKER_RUN:=
 else
-	DOCKER_RUN:=docker run --rm -it \
-		-v "`pwd`:`pwd`" \
-		-v "${COMPOSER_CACHE_DIR}:${COMPOSER_CONTAINER_CACHE_DIR}" \
-		-w "`pwd`" \
-		-e OTEL_PHP_FIBERS_ENABLED="true" \
-		"${CONTAINER_NAME}"
+    ifeq ($(C_EXIT_STATUS),0)
+        DOCKER_RUN:=docker run --rm -it \
+            -v "`pwd`:`pwd`" \
+            -v "${COMPOSER_CACHE_DIR}:${COMPOSER_CONTAINER_CACHE_DIR}" \
+            -w "`pwd`" \
+            -e OTEL_PHP_FIBERS_ENABLED="true" \
+            "${CONTAINER_NAME}"
+    else
+        DOCKER_RUN:=
+    endif
 endif
 
 ifneq (,$(findstring icrosoft,$(shell cat /proc/version)))
